Setting DETECT_CONNECTED_DEVICE
This is required because of a JRuby bug where JRuby can't kill
subprocesses running in a Timeout block. For more information see -
https://github.com/navaneeth/calabash-ios/commit/4f5ca1f912198dbe705e10047060820e26378159

From the commit on calabash-ios:

"Some Ruby implementation like JRuby has trouble executing a process in
the Timeout::timeout(). JRuby won't throw the exception and hangs until
the process returns.

In Calabash launcher, it detects the connected devices by executing the
external process: udidetect. udidetect hangs when there are no devices
connected. Calabsh launcher executes udidetect in a Timeout and when it
timesout, calabash kills the udidetect process.

In Jruby, because of a JRuby bug, this causes the process to wait until
the launched process exits. TimeoutError is never fired.

This commit introduces an environment variable DETECT_CONNECTED_DEVICE
which can be used to turn on/off device detection. By default device
detection is on."